Cardapio e outras coisas finalizadas

+++ database/migrations/2020_03_18_231547_create_restaurants_table.php
@@ -22,9 +22,9 @@ class CreateRestaurantsTable extends Migration
             $table->integer('cep')->nullable();
             $table->bigInteger('telefone')->nullable();
             $table->biginteger('celular');
-            $table->double('vlr_m')->nullable();
-            $table->double('frete')->nullable();
-            $table->double('adicional')->nullable();
+            $table->decimal('vlr_m',5,2)->nullable();
+            $table->decimal('frete',5,2)->nullable();
+            $table->decimal('adicional',5,2)->nullable();
             $table->string('responsavel');
             $table->boolean('cobfr')->nullable(); // Cobra Frete ?
             $table->boolean('cobad')->nullable(); // Cobra Adicional ?
